backend/simulations/oasis/requirements.txt
camel-ai>=0.2.78
camel-oasis>=0.2.5
openai
